+++ examples/rtu_broadcast/broadcast_slave.c
@@ -97,12 +97,13 @@ static void print_progress(size_t cur_size, size_t total_size)
  * @brief   从机回调函数
  * @param   ctx modbus 句柄
  * @param   slave_info 从机信息体
+ * @param   data 私有数据
  * @return  =0:正常;
  *          <0:异常
  *             (-AGILE_MODBUS_EXCEPTION_UNKNOW(-255): 未知异常，从机不会打包响应数据)
  *             (其他负数异常码: 从机会打包异常响应数据)
  */
-static int slave_callback(agile_modbus_t *ctx, struct agile_modbus_slave_info *slave_info)
+static int slave_callback(agile_modbus_t *ctx, struct agile_modbus_slave_info *slave_info, const void *data)
 {
     int function = slave_info->sft->function;
 
@@ -297,7 +298,7 @@ static void *cycle_entry(void *param)
 
         // 解包，为了防止脏数据，不能直接丢，往后挪一个字节继续解析
         while (total_len > 0) {
-            int rc = agile_modbus_slave_handle(ctx, total_len, 1, slave_callback, &frame_length);
+            int rc = agile_modbus_slave_handle(ctx, total_len, 1, slave_callback, NULL, &frame_length);
             if (rc >= 0) {
                 ctx->read_buf = ctx->read_buf + frame_length;
                 ctx->read_bufsz = ctx->read_bufsz - frame_length;