+++ src/agile_modbus.c
@@ -2,7 +2,7 @@
  * @file    agile_modbus.c
  * @brief   Agile Modbus 软件包通用源文件
  * @author  马龙伟 (2544047213@qq.com)
- * @date    2021-12-02
+ * @date    2022-07-28
  *
  @verbatim
     使用：
@@ -1196,11 +1196,12 @@ uint16_t agile_modbus_slave_register_get(uint8_t *buf, int index)
  *     @arg 0: 不比对从机地址
  *     @arg 1: 比对从机地址
  * @param   slave_cb 从机回调函数
+ * @param   slave_data 从机回调函数私有数据
  * @param   frame_length 存放 modbus 数据帧长度
  * @return  >=0:要响应的数据长度; 其他:异常
  */
 int agile_modbus_slave_handle(agile_modbus_t *ctx, int msg_length, uint8_t slave_strict,
-                              agile_modbus_slave_callback_t slave_cb, int *frame_length)
+                              agile_modbus_slave_callback_t slave_cb, const void *slave_data, int *frame_length)
 {
     int min_rsp_length = ctx->backend->header_length + 5 + ctx->backend->checksum_length;
     if (ctx->send_bufsz < min_rsp_length)
@@ -1475,7 +1476,7 @@ int agile_modbus_slave_handle(agile_modbus_t *ctx, int msg_length, uint8_t slave
         rsp_length = agile_modbus_serialize_response_exception(ctx, &sft, exception_code);
     else {
         if (slave_cb) {
-            int ret = slave_cb(ctx, &slave_info);
+            int ret = slave_cb(ctx, &slave_info, slave_data);
 
             if (ret < 0) {
                 if (ret == -AGILE_MODBUS_EXCEPTION_UNKNOW)